+++ content/webapp/components/WorkDetails/WorkDetails.Holdings.tsx
@@ -1,0 +1,78 @@
+import { font } from '@weco/common/utils/classnames';
+import ExpandableList from '@weco/common/views/components/ExpandableList/ExpandableList';
+import Space from '@weco/common/views/components/styled/Space';
+import WorkDetailsSection from './WorkDetails.Section';
+import WorkDetailsText from './WorkDetails.Text';
+import {
+  getLocationLabel,
+  getLocationLink,
+  getLocationShelfmark,
+} from '@weco/content/utils/works';
+import { Holding } from '@weco/content/services/wellcome/catalogue/types';
+
+const WorkDetailsHoldings = ({ holdings }: { holdings: Holding[] }) => {
+  return (
+    <>
+      {holdings.length > 0 ? (
+        <WorkDetailsSection headingText="Holdings">
+          {holdings.map((holding, i) => {
+            const locationLabel =
+              holding.location && getLocationLabel(holding.location);
+            const locationShelfmark =
+              holding.location && getLocationShelfmark(holding.location);
+            const locationLink =
+              holding.location && getLocationLink(holding.location);
+            return (
+              <Space
+                key={i}
+                v={
+                  i + 1 !== holdings.length
+                    ? { size: 'l', properties: ['margin-bottom'] }
+                    : { size: 'l', properties: [] }
+                }
+              >
+                {holding.enumeration.length > 0 && (
+                  <Space
+                    key={i}
+                    v={{ size: 's', properties: ['margin-bottom'] }}
+                  >
+                    <ExpandableList
+                      listItems={holding.enumeration}
+                      initialItems={10}
+                    />
+                  </Space>
+                )}
+                <Space key={i} v={{ size: 's', properties: ['margin-bottom'] }}>
+                  {locationLink && (
+                    <a className={font('intr', 5)} href={locationLink.url}>
+                      {locationLink.linkText}
+                    </a>
+                  )}
+
+                  {locationShelfmark && (
+                    <WorkDetailsText
+                      title="Location"
+                      noSpacing={true}
+                      text={`${locationLabel} ${locationShelfmark}`}
+                    />
+                  )}
+
+                  {holding.note && (
+                    <WorkDetailsText
+                      title="Note"
+                      inlineHeading={true}
+                      noSpacing={true}
+                      text={holding.note}
+                    />
+                  )}
+                </Space>
+              </Space>
+            );
+          })}
+        </WorkDetailsSection>
+      ) : null}
+    </>
+  );
+};
+
+export default WorkDetailsHoldings;
