+++ content/webapp/components/WorkDetails/WorkDetails.List.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { FunctionComponent } from 'react';
 
 import Space from '@weco/common/views/components/styled/Space';
-import WorkDetailsProperty from '../WorkDetailsProperty/WorkDetailsProperty';
+import WorkDetailsProperty from './WorkDetails.Property';
 
 const PlainList = styled(Space)`
   list-style: none;
