+++ content/webapp/components/WorkDetails/WorkDetails.Property.tsx
@@ -1,9 +1,9 @@
+import { FunctionComponent, PropsWithChildren } from 'react';
+import styled from 'styled-components';
 import { font } from '@weco/common/utils/classnames';
 import SpacingComponent from '@weco/common/views/components/styled/SpacingComponent';
 import Space from '@weco/common/views/components/styled/Space';
 import ConditionalWrapper from '@weco/common/views/components/ConditionalWrapper/ConditionalWrapper';
-import { FunctionComponent, PropsWithChildren } from 'react';
-import styled from 'styled-components';
 
 type InlineHeadingProps = {
   inlineHeading?: boolean;
