+++ content/webapp/components/WorkDetails/WorkDetails.Tags.tsx
@@ -1,8 +1,8 @@
+import { FunctionComponent } from 'react';
 import Tags, {
   Props as TagsProps,
 } from '@weco/common/views/components/Tags/Tags';
-import WorkDetailsProperty from '../WorkDetailsProperty/WorkDetailsProperty';
-import { FunctionComponent } from 'react';
+import WorkDetailsProperty from './WorkDetails.Property';
 import Space from '@weco/common/views/components/styled/Space';
 import ConditionalWrapper from '@weco/common/views/components/ConditionalWrapper/ConditionalWrapper';
