+++ content/webapp/components/WorkDetails/WorkDetails.Text.tsx
@@ -1,9 +1,9 @@
+import styled from 'styled-components';
+import { FunctionComponent, ReactElement } from 'react';
 import WorkDetailsProperty, {
   Props as BaseProps,
-} from '../WorkDetailsProperty/WorkDetailsProperty';
-import { FunctionComponent, ReactElement } from 'react';
+} from './WorkDetails.Property';
 import { isString } from '@weco/common/utils/type-guards';
-import styled from 'styled-components';
 
 const LimitWidth = styled.div.attrs({
   className: 'spaced-text',
